refactor: 위치와 관련된 클래스 패키지 이동
