+++ src/main/java/model/piece/Jol.java
@@ -1,10 +1,12 @@
 package model.piece;
 
+import model.piece.position.Position;
+
 import java.util.List;
 
-import static model.piece.MovementChecker.isUp;
-import static model.piece.MovementChecker.isLeft;
-import static model.piece.MovementChecker.isRight;
+import static model.piece.position.MovementChecker.isUp;
+import static model.piece.position.MovementChecker.isLeft;
+import static model.piece.position.MovementChecker.isRight;
 
 public class Jol extends Piece {
