+++ src/main/java/model/piece/Piece.java
@@ -1,5 +1,7 @@
 package model.piece;
 
+import model.piece.position.Position;
+
 import java.util.List;
 import java.util.Objects;
