+++ src/main/java/model/piece/position/MovementChecker.java
@@ -1,4 +1,4 @@
-package model.piece;
+package model.piece.position;
 
 public class MovementChecker {
