+++ src/main/java/model/piece/position/Position.java
@@ -1,4 +1,4 @@
-package model.piece;
+package model.piece.position;
 
 import java.util.Objects;
