+++ src/test/java/model/piece/ByeongTest.java
@@ -1,5 +1,6 @@
 package model.piece;
 
+import model.piece.position.Position;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
